normal commit

+++ openmp/inc/functions.h
@@ -18,7 +18,7 @@ void testScatter_intr(int *source, int *dest, int* loc, int n);
 void testScan_tbb(int* a, int* b, int n, int pattern);
 double testScan_omp(int *a, int *b, int n, int pattern);
 
-double map(float *source, float *dest, int n);
+double map(int *source, int *dest, int n);
 
 double testScan_ass(int* a, int* b, int n, int pattern);
